+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - New types defined in `VocdoniCensus3Client` with some additional documentation.
+- Added new field `symbol` to summary token list in Census3.
 
 ### Fixed